# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orange Banking API",
        "description": "API documentation for Orange Banking System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/transactions": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "description": "Create a new transaction on one of the client's accounts",
                "operationId": "createTransaction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compte_id",
                                    "type",
                                    "montant"
                                ],
                                "properties": {
                                    "compte_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "f9136320-10c8-47cc-8de0-ba4cc2ef78b1"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "depot",
                                            "retrait",
                                            "transfert",
                                            "paiement"
                                        ],
                                        "example": "depot"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "100.50"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Dépôt d'argent"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction créée"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies sont invalides."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès au compte refusé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Transaction": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "39adc2af-39c8-41cf-8aa7-930ba114aad0"
                    },
                    "compte_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "f9136320-10c8-47cc-8de0-ba4cc2ef78b1"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "reception",
                            "transfert",
                            "paiement"
                        ],
                        "example": "reception"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": "100.50"
                    },
                    "description": {
                        "type": "string",
                        "example": "Dépôt d'argent"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}